+++ live_table.py
@@ -1,5 +1,5 @@
 import argparse
-from animated_multi_player_game import run_simulation_with_animation
+from multi_player_game import GameSimulation
 
 def main():
     # Create the parser
@@ -11,8 +11,8 @@ def main():
     parser.add_argument('-folder', type=str, required=False, help='folder for which classes to run')
     # Parse arguments
     args = parser.parse_args()
-
-    run_simulation_with_animation(number = args.sims, refresh_rate = args.refresh, folder_name = args.folder)
+    simulation = GameSimulation
+    simulation.run_simulation_with_animation(number = args.sims, refresh_rate = args.refresh, folder_name = args.folder)
 
     # Run the simulation
     #results = run_simulation_many_times(args.sims)
